added some section in the contact-us page
